pyproject.toml
[project]
name = "medihelp-redesigned"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular>=0.28.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "radis>=0.16.2",
    "requests>=2.32.4",
]
